+++ test/spec/multibrowser/instance/execute.js
@@ -1,9 +1,11 @@
 /* global document */
-describe('execute', function() {
-    before(h.setup());
+var async = require('async');
+
+describe('execute executed by single multibrowser instance', function() {
+    before(h.setupMultibrowser());
 
     it('should be able to execute some js', function(done) {
-        this.client
+        this.browserA
             .execute('return document.title', [], function(err, res) {
                 assert.ifError(err);
                 assert.equal(conf.testPage.title, res.value);
@@ -12,7 +14,7 @@ describe('execute', function() {
     });
 
     it('should be forgiving on giving an `args` parameter', function(done) {
-        this.client
+        this.browserA
             .execute('return document.title', function(err, res) {
                 assert.ifError(err);
                 assert.equal(conf.testPage.title, res.value);
@@ -21,7 +23,7 @@ describe('execute', function() {
     });
 
     it('should be able to execute a pure function', function(done) {
-        this.client
+        this.browserA
             .execute(function() {
                 return document.title;
             }, function(err, res) {
@@ -29,20 +31,6 @@ describe('execute', function() {
                 assert.equal(conf.testPage.title, res.value);
                 done(err);
             });
-    });
-
-    it('should be able to take just a single function', function(done) {
-        this.client
-            .execute(function() {
-                window.testThatStuff = true;
-            })
-            .execute(function() {
-                return window.testThatStuff;
-            }, function(err, res) {
-                assert.ifError(err);
-                assert.equal(res.value, true);
-            })
-            .call(done);
     });
 
 });