+++ test/spec/multibrowser/instance/promises.js
@@ -1,12 +1,11 @@
-var Q = require('q');
+describe('PromiseHandler executed by single multibrowser instance', function() {
 
-describe('PromiseHandler', function() {
-
-    before(h.setup());
+    before(h.setupMultibrowser());
 
     it('should sync promises with call', function(done) {
         var result = '';
-        this.client
+        this.browserA
+            .call() // WebdriverIO instances need to call at least one command to enable promises
             .then(function() {
                 result += '1';
             })
@@ -26,7 +25,7 @@ describe('PromiseHandler', function() {
     });
 
     it('should propagate results to then', function(done) {
-        this.client
+        this.browserA
             .getTitle().then(function(title) {
                 title.should.be.equal('WebdriverJS Testpage');
                 return this.url();
@@ -46,11 +45,11 @@ describe('PromiseHandler', function() {
     it('should be working on custom commands', function(done) {
         var result = '';
 
-        this.client.addCommand('fakeCommand', function(param, done) {
+        this.browserA.addCommand('fakeCommand', function(param, done) {
             done(undefined, param);
         });
 
-        this.client
+        this.browserA
             .fakeCommand(0)
             .then(function() {
                 return this.fakeCommand(1);
@@ -79,7 +78,7 @@ describe('PromiseHandler', function() {
 
     it('should reject promise if command throws an error', function(done) {
         var result = null;
-        this.client
+        this.browserA
             .click('#notExisting').then(function() {
                 result = false;
             }, function() {
@@ -92,7 +91,7 @@ describe('PromiseHandler', function() {
     });
 
     it('should handle waitfor commands within then callbacks', function(done) {
-        this.client
+        this.browserA
             .getTitle().then(function() {
                 return this.pause(1000).pause(100).isVisible('body');
             }).then(function(result) {
@@ -105,7 +104,7 @@ describe('PromiseHandler', function() {
         var gotExecutedCatch = false,
             gotExecutedFail = false;
 
-        this.client
+        this.browserA
             .click('#notExisting').catch(function() {
                 gotExecutedCatch = true;
             })
@@ -125,7 +124,7 @@ describe('PromiseHandler', function() {
         var gotExecutedCatch = false,
             gotExecutedFail = false;
 
-        this.client
+        this.browserA
             .click('body').catch(function() {
                 gotExecutedCatch = true;
             })
@@ -143,7 +142,7 @@ describe('PromiseHandler', function() {
 
     it('should propagate not only promises but also objects or strings', function(done) {
         var hasBeenExecuted = 0;
-        this.client
+        this.browserA
             .isVisible('body').then(function(isVisible) {
                 hasBeenExecuted++;
                 return isVisible;
@@ -164,36 +163,6 @@ describe('PromiseHandler', function() {
                 hasBeenExecuted.should.be.equal(4);
                 done();
             });
-    });
-
-    describe('should be able to handle 3rd party promises', function() {
-
-        it('should handle Q\'s deferred.promise', function(done) {
-
-            var deferred = Q.defer();
-            deferred.resolve('success');
-
-            this.client
-                .status()
-                .then(function(){
-                    return deferred.promise;
-                })
-                .then(function(result){
-                    result.should.be.equal('success');
-                })
-                .call(done);
-
-        });
-
-    });
-
-    it('should be able to pass a command execution as parameter', function(done) {
-        this.client
-            .then(this.client.getTitle())
-            .then(function(title) {
-                title.should.be.exactly(conf.testPage.title);
-            })
-            .call(done);
     });
 
 });